Move servlet adapters to an internal package to avoid duplicating classes when building OSGi bundles
